requirements.txt
datasets>=2.19.0
google-genai>=1.30.0
python-dotenv>=1.0.1
requests>=2.32.0
tqdm>=4.66.0
